FEATURE: start screen added
