--- f20_map_quizgame/answerset.cpp
@@ -1,11 +1,0 @@
-#include "answerset.h"
-
-AnswerSet::AnswerSet()
-{
-
-}
-
-void AnswerSet::addAnswer(QString ans)
-{
-    answers.push_back(ans);
-}
