+++ f20_map_quizgame/logic/quiz.cpp
@@ -1,0 +1,6 @@
+#include "quiz.h"
+
+Quiz::Quiz()
+{
+
+}
