+++ f20_map_quizgame/logic/quiz.h
@@ -1,0 +1,16 @@
+#ifndef QUIZ_H
+#define QUIZ_H
+
+#include<vector>
+
+class Quiz
+{
+public:
+    Quiz();
+
+
+private:
+    unsigned int currentQuestion = 0;
+};
+
+#endif // QUIZ_H
