+++ f20_map_quizgame/main.cpp
@@ -1,11 +1,11 @@
-#include "mainwindow.h"
+#include "ui/mainwindow.h"
 
 #include <QApplication>
 
 int main(int argc, char *argv[])
 {
     QApplication a(argc, argv);
-    MainWindow w;
-    w.show();
+    MainWindow screen;
+    screen.show();
     return a.exec();
 }
