+++ f20_map_quizgame/ui/mainwindow.h
@@ -2,8 +2,6 @@
 #define MAINWINDOW_H
 
 #include <QMainWindow>
-#include <vector>
-#include "answerset.h"
 
 
 QT_BEGIN_NAMESPACE
@@ -19,12 +17,11 @@ public:
     ~MainWindow();
 
 private slots:
-    void on_nextButton_clicked();
+    void on_button_to_quiz_list_clicked();
+
+    void on_button_attempt_quiz_clicked();
 
 private:
     Ui::MainWindow *ui;
-    std::vector<QString> questions;
-    std::vector<AnswerSet> a_set;
-    unsigned int current_question = 0;
 };
 #endif // MAINWINDOW_H
